[TASK] update cmake configure for iphone-simulate project. remove warning of code

+++ libguiex_module/stringconv_iconv/guistringconv_iconv.cpp
@@ -82,7 +82,7 @@ namespace guiex
 			} 
 
 			//check ret
-			if( retbytes == -1 )
+			if( retbytes == size_t(-1) )
 			{
 				if( errno == E2BIG )
 				{
@@ -178,7 +178,7 @@ namespace guiex
 			} 
 
 			//check ret
-			if( retbytes == -1 )
+			if( retbytes == size_t(-1) )
 			{
 				if( errno == E2BIG )
 				{
